+++ src/pymgrid/microgrid/modules/__init__.py
@@ -1,7 +1,6 @@
-from src.pymgrid.microgrid.modules.base_module import BaseMicrogridModule, BaseTimeSeriesMicrogridModule
-from src.pymgrid.microgrid.modules.battery_module import BatteryModule
-from src.pymgrid.microgrid.modules.renewable_module import RenewableModule
-from src.pymgrid.microgrid.modules.load_module import LoadModule
-from src.pymgrid.microgrid.modules.unbalanced_energy_module import UnbalancedEnergyModule
-from src.pymgrid.microgrid.modules.grid_module import GridModule
-from src.pymgrid.microgrid.modules.genset_module import GensetModule
+from .battery_module import BatteryModule
+from .renewable_module import RenewableModule
+from .load_module import LoadModule
+from .unbalanced_energy_module import UnbalancedEnergyModule
+from .grid_module import GridModule
+from .genset_module import GensetModule
