+++ src/pymgrid/microgrid/modules/base/__init__.py
@@ -1,0 +1,2 @@
+from pymgrid.microgrid.modules.base.base_module import BaseMicrogridModule
+from pymgrid.microgrid.modules.base.base_timeseries_module import BaseTimeSeriesMicrogridModule
