+++ src/pymgrid/microgrid/modules/battery_module.py
@@ -1,4 +1,4 @@
-from src.pymgrid.microgrid.modules.base_module import BaseMicrogridModule
+from pymgrid.microgrid.modules.base import BaseMicrogridModule
 import numpy as np
